[z2082] Fix more warnings
* Remove warning: enumeration value '<value>' not handled in switch
* Remove warning: enumeral and non-enumeral type in conditional expression
* Remove warning: too many arguments for format


(based on commit [12121] - 8209548)

(based on commit [s1700] - 1c6fc40)

+++ src/game/BattleGround/BattleGround.cpp
@@ -71,7 +71,7 @@ namespace MaNGOS
                 data << ObjectGuid(targetGuid);             // there 0 for BG messages
                 data << uint32(strlen(text) + 1);
                 data << text;
-                data << uint8(i_source ? i_source->GetChatTag() : CHAT_TAG_NONE);
+                data << uint8(i_source ? i_source->GetChatTag() : uint8(CHAT_TAG_NONE));
             }
 
             ChatMsg i_msgtype;
@@ -147,7 +147,7 @@ namespace MaNGOS
                 data << ObjectGuid(targetGuid);             // there 0 for BG messages
                 data << uint32(strlen(str) + 1);
                 data << str;
-                data << uint8(i_source ? i_source->GetChatTag() : CHAT_TAG_NONE);
+                data << uint8(i_source ? i_source->GetChatTag() : uint8(CHAT_TAG_NONE));
             }
         private:
